+++ apps/www/next.config.js
@@ -3,6 +3,11 @@ const { createSecurityHeaderRoutes } = require("@phylopic/ui/securityHeaders")
 /** @type {import('next').NextConfig} */
 const nextConfig = {
     outputFileTracingRoot: path.join(__dirname, "../../"),
+    serverExternalPackages: [
+        "@aws-sdk/credential-provider-web-identity",
+        "@vercel/functions",
+        "@vercel/oidc",
+    ],
     async headers() {
         return [
             ...createSecurityHeaderRoutes(),